Fix profile recheck error (#4690)
