+++ apm-checkstyle/checkStyle.xml
@@ -96,7 +96,7 @@
             <property name="format" value="(^[A-Z][A-Z0-9]*(_[A-Z0-9]+)*$)"/>
         </module>
         <module name="PackageName">
-            <property name="format" value="^org\.apache\.skywalking(\.[a-zA-Z][a-zA-Z0-9]*)+$"/>
+            <property name="format" value="^(org|test)\.apache\.skywalking(\.[a-zA-Z][a-zA-Z0-9]*)+$"/>
         </module>
         <module name="ParameterName"/>
         <module name="StaticVariableName">
