+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/context/tag/Tags.java
@@ -80,7 +80,7 @@ public final class Tags {
     public static final class HTTP {
         public static final StringTag METHOD = new StringTag(10, "http.method");
 
-        public static final StringTag PARAMS = new StringTag(11, "http.params");
+        public static final StringTag PARAMS = new StringTag(11, "http.params", true);
     }
 
     /**
