+++ test/e2e/e2e-test/src/test/resources/expected/profile/profileSegments.yml
@@ -16,7 +16,7 @@
 traces:
   - key: not null
     endpointNames:
-      - /profile/users
+      - /profile/{name}
     duration: ge 0
     start: gt 0
     isError: false
